feat(src): Implement Naive Bayes model
A Naive Bayes model has been completed that features allows for
modularity in the settings. The main.py script is the driver of the
application, and allows teh user to determine the datasets being used
and the query to be searched via CLI prompts and flags.

Certain features still remain to be implemented, such as a Gaussian
model option, a flag to toggle Laplacian smoothing, and more
importantly, a full Bayesian model.